Create distinct `AsyncEventResult` type
The difference is that an `EventResult` can carry an updated state, while an
`AsyncEventResult` cannot carry a state, but can carry a function that produces
an updated state given the current state.

Between the time when a promise for an `AsyncEventResult` is returned from
a reducer and when the promise resolves, any number of changes could have been
made to app state. Replacing state without regard to those changes is something
that should probably never be done. That is why `AsyncEventResult` carries an
updater function.